requirements.txt
# 데이터 수집
feedparser==6.0.10
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# AI/ML
google-generativeai==0.3.0
anthropic==0.7.0
transformers==4.35.0
torch==2.8.0
sentence-transformers==2.2.2

# 이메일 & 스케줄링
schedule==1.2.0
pytz==2023.3
jinja2==3.1.2

# Google API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# 유틸리티
python-dotenv==1.0.0